requirements.txt
# Python Requirements for Photonic Lantern Digital Holography Automation
# 
# Installation:
#   pip install -r requirements.txt
#
# Note: Some packages require special installation (see README.md):
#   - Xeneth SDK: Install from "C:\Program Files\Xeneth\Sdk\Python\..."
#   - Thorlabs Kinesis: Download from Thorlabs website

# Core scientific packages
numpy>=1.20.0
scipy>=1.7.0
matplotlib>=3.4.0

# Configuration and data
pyyaml>=5.4.0

# Hardware communication
pyserial>=3.5
pyvisa>=1.11.0
pyvisa-py>=0.5.0

# Image processing
opencv-python>=4.5.0
Pillow>=8.0.0

# Xeneth camera SDK (must be installed separately)
# Install after installing Xeneth software:
# pip install "C:\Program Files\Xeneth\Sdk\Python\xenics-3.x.x-py3-none-any.whl"

# Optional: For enhanced functionality
# ipython>=7.0.0  # Interactive shell
# jupyter>=1.0.0  # Jupyter notebooks
# pandas>=1.3.0   # Data analysis
# tqdm>=4.60.0    # Progress bars
